requirements.txt
torch==2.1.0
backoff==2.2.1
bitsandbytes==0.39.0
bert-score==0.3.13
jsonlines==3.1.0
pyenchant==3.2.2
nltk==3.7
datasets==2.16.1
deepspeed=0.12.6
flash-attn==2.4.2
fschat==0.2.35
openai==1.7.2
peft==0.7.1
sentencepiece==0.1.99
spacy==3.7.2
scipy==1.11.4
tqdm==4.66.1
transformers==4.33.0
